feat: refactor the token cleanup logic

+++ src/IdentityServer4.EntityFramework/TokenCleanup.cs
@@ -7,36 +7,29 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using IdentityServer4.EntityFramework.DbContexts;
-using IdentityServer4.EntityFramework.Interfaces;
 using IdentityServer4.EntityFramework.Options;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.DependencyInjection;
 
-namespace IdentityServer4.EntityFramework
-{
-    public class TokenCleanupOptions
-    {
-        public DbContextOptions<PersistedGrantDbContext> DbContextOptions { get; set; }
-        public ILoggerFactory LoggerFactory { get; set; }
-        public int Interval { get; set; } = 60;
-    }
-
+namespace IdentityServer4.EntityFramework {
     public class TokenCleanup
     {
-        private readonly DbContextOptions<PersistedGrantDbContext> _options;
         private readonly TimeSpan _interval;
         private CancellationTokenSource _source;
         private readonly ILogger _logger;
+        private readonly IServiceProvider _serviceProvider;
 
-        public TokenCleanup(TokenCleanupOptions options)
+        // The logger cross-thread, so ...
+        //public TokenCleanup(OperationalStoreOptions options, ILogger<TokenCleanup> logger, IServiceProvider serviceProvider)
+        public TokenCleanup(OperationalStoreOptions options, IServiceProvider serviceProvider)
         {
             if (options == null) throw new ArgumentNullException(nameof(options));
-            if (options.DbContextOptions == null) throw new ArgumentNullException(nameof(options.DbContextOptions));
-            if (options.Interval < 1) throw new ArgumentException("interval must be more than 1 second");
+            if (options.TokenCleanup == null) throw new ArgumentNullException(nameof(options.TokenCleanup));
+            if (options.TokenCleanup.Interval < 1) throw new ArgumentException("interval must be more than 1 second");
 
-            _options = options.DbContextOptions;
-            _interval = TimeSpan.FromSeconds(options.Interval);
-            _logger = options.LoggerFactory?.CreateLogger(typeof(TokenCleanup).FullName) ?? new NopLogger();
+            _serviceProvider = serviceProvider;
+            _interval = TimeSpan.FromSeconds(options.TokenCleanup.Interval);
+            _logger = new NopLogger();
         }
 
         public void Start()
@@ -89,27 +82,28 @@ namespace IdentityServer4.EntityFramework
             }
         }
 
-        protected virtual IPersistedGrantDbContext CreateOperationalDbContext()
-        {
-            // TODO: Replace with DI.
-            return new PersistedGrantDbContext(_options, new OperationalStoreOptions());
-        }
-
         private void ClearTokens()
         {
             try
             {
                 _logger.LogTrace("Querying for tokens to clear");
 
-                using (var context = CreateOperationalDbContext())
+                // Because the PersistedGrantDbContext is scoped lifetime.
+                using (var serviceScope = _serviceProvider.GetRequiredService<IServiceScopeFactory>().CreateScope())
                 {
-                    var expired = context.PersistedGrants.Where(x => x.Expiration < DateTimeOffset.UtcNow).ToArray();
+                    using (var context = serviceScope.ServiceProvider.GetService<PersistedGrantDbContext>())
+                    {
+                        var expired = context.PersistedGrants.Where(x => x.Expiration < DateTimeOffset.UtcNow).ToArray();
 
-                    _logger.LogDebug("Clearing {tokenCount} tokens", expired.Length);
+                        _logger.LogDebug("Clearing {tokenCount} tokens", expired.Length);
 
-                    context.PersistedGrants.RemoveRange(expired);
+                        if (expired.Length > 0) 
+                        {
+                            context.PersistedGrants.RemoveRange(expired);
 
-                    context.SaveChanges();
+                            context.SaveChanges();
+                        }
+                    }
                 }
             }
             catch (Exception ex)
